fix: split account info from credentials
