+++ cosmic-accounts-daemon/src/models/mod.rs
@@ -1,0 +1,2 @@
+mod provider;
+pub use provider::{AccountProviderConfig, ProviderConfig};
