+++ cosmic-accounts-gui/src/main.rs
@@ -2,7 +2,8 @@ use cosmic::app::Task;
 use cosmic::iced::{Alignment, Length};
 use cosmic::widget::image::Handle;
 use cosmic::{widget, Application, Core, Element};
-use cosmic_accounts::{Account, CosmicAccountsClient, Provider, Uuid};
+use cosmic_accounts::models::{Account, Provider};
+use cosmic_accounts::{CosmicAccountsClient, Uuid};
 use tracing::info;
 use tracing_subscriber::EnvFilter;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
