+++ src/client.rs
@@ -1,7 +1,10 @@
 use uuid::Uuid;
 use zbus::Connection;
 
-use crate::{proxy::CosmicAccountsProxy, Account, Provider, Result};
+use crate::{
+    models::{Account, Provider},
+    proxy::CosmicAccountsProxy,
+};
 
 #[derive(Debug, Clone)]
 pub struct CosmicAccountsClient {
@@ -9,7 +12,7 @@ pub struct CosmicAccountsClient {
 }
 
 impl<'a> CosmicAccountsClient {
-    pub async fn new() -> Result<Self> {
+    pub async fn new() -> Result<Self, zbus::fdo::Error> {
         let connection = Connection::session().await?;
         let proxy = CosmicAccountsProxy::new(&connection).await?;
         Ok(Self { proxy })
@@ -17,45 +20,35 @@ impl<'a> CosmicAccountsClient {
 }
 
 impl CosmicAccountsClient {
-    pub async fn list_accounts(&self) -> Result<Vec<Account>> {
+    pub async fn list_accounts(&self) -> Result<Vec<Account>, zbus::fdo::Error> {
         self.proxy
             .list_accounts()
             .await
             .map(|accounts| accounts.into_iter().map(Into::into).collect())
-            .map_err(Into::into)
     }
 
-    pub async fn start_authentication(&mut self, provider: &Provider) -> Result<String> {
-        self.proxy
-            .start_authentication(&provider.to_string())
-            .await
-            .map_err(Into::into)
+    pub async fn start_authentication(
+        &mut self,
+        provider: &Provider,
+    ) -> Result<String, zbus::fdo::Error> {
+        self.proxy.start_authentication(&provider.to_string()).await
     }
 
     pub async fn complete_authentication(
         &mut self,
         csrf_token: &str,
         authorization_code: &str,
-    ) -> Result<String> {
+    ) -> Result<String, zbus::fdo::Error> {
         self.proxy
             .complete_authentication(csrf_token, authorization_code)
             .await
-            .map_err(Into::into)
     }
 
-    pub async fn get_account(&self, id: &str) -> Result<Account> {
-        self.proxy
-            .get_account(id)
-            .await
-            .map(Into::into)
-            .map_err(Into::into)
+    pub async fn get_account(&self, id: &str) -> Result<Account, zbus::fdo::Error> {
+        self.proxy.get_account(id).await.map(Into::into)
     }
 
-    pub async fn remove_account(&mut self, id: &Uuid) -> Result<()> {
-        self.proxy
-            .remove_account(&id.to_string())
-            .await
-            .map(Into::into)
-            .map_err(Into::into)
+    pub async fn remove_account(&mut self, id: &Uuid) -> Result<(), zbus::fdo::Error> {
+        self.proxy.remove_account(&id.to_string()).await
     }
 }
