+++ src/models/account.rs
@@ -1,8 +1,29 @@
+use std::str::FromStr;
+
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+
+use crate::models::{Capability, Provider};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct Account {
+    pub id: Uuid,
+    pub provider: Provider,
+    pub display_name: String,
+    pub username: String,
+    pub email: Option<String>,
+    pub enabled: bool,
+    pub created_at: DateTime<Utc>,
+    pub last_used: Option<DateTime<Utc>>,
+    pub capabilities: Vec<Capability>,
+}
+
 use zbus::zvariant::{DeserializeDict, SerializeDict, Type};
 
 #[derive(Debug, Clone, PartialEq, DeserializeDict, SerializeDict, Type)]
 #[zvariant(signature = "dict")]
-pub struct Account {
+pub struct DbusAccount {
     pub id: String,
     pub provider: String,
     pub display_name: String,
@@ -11,22 +32,11 @@ pub struct Account {
     pub enabled: bool,
     pub created_at: String,
     pub last_used: Option<String>,
-    pub credentials: Credentials,
     pub capabilities: Vec<String>,
 }
 
-#[derive(Debug, Clone, PartialEq, DeserializeDict, SerializeDict, Type)]
-#[zvariant(signature = "dict")]
-pub struct Credentials {
-    pub access_token: String,
-    pub refresh_token: Option<String>,
-    pub expires_at: Option<String>,
-    pub scope: Vec<String>,
-    pub token_type: String,
-}
-
-impl From<super::Account> for Account {
-    fn from(value: super::Account) -> Self {
+impl From<Account> for DbusAccount {
+    fn from(value: Account) -> Self {
         Self {
             id: value.id.to_string(),
             provider: value.provider.to_string(),
@@ -39,16 +49,6 @@ impl From<super::Account> for Account {
                 .last_used
                 .clone()
                 .map(|last_used| last_used.to_string()),
-            credentials: Credentials {
-                access_token: value.credentials.access_token,
-                refresh_token: value.credentials.refresh_token,
-                expires_at: value
-                    .credentials
-                    .expires_at
-                    .map(|expires_at| expires_at.to_string()),
-                scope: value.credentials.scope,
-                token_type: value.credentials.token_type,
-            },
             capabilities: value
                 .capabilities
                 .iter()
@@ -58,8 +58,8 @@ impl From<super::Account> for Account {
     }
 }
 
-impl From<&super::Account> for Account {
-    fn from(value: &super::Account) -> Self {
+impl From<&Account> for DbusAccount {
+    fn from(value: &Account) -> Self {
         Self {
             id: value.id.to_string(),
             provider: value.provider.to_string(),
@@ -72,21 +72,30 @@ impl From<&super::Account> for Account {
                 .last_used
                 .clone()
                 .map(|last_used| last_used.to_string()),
-            credentials: Credentials {
-                access_token: value.credentials.access_token.clone(),
-                refresh_token: value.credentials.refresh_token.clone(),
-                expires_at: value
-                    .credentials
-                    .expires_at
-                    .map(|expires_at| expires_at.to_string()),
-                scope: value.credentials.scope.clone(),
-                token_type: value.credentials.token_type.clone(),
-            },
             capabilities: value
                 .capabilities
                 .iter()
                 .map(|capability| capability.to_string())
                 .collect(),
+        }
+    }
+}
+
+impl From<DbusAccount> for Account {
+    fn from(value: DbusAccount) -> Self {
+        Account {
+            id: Uuid::from_str(&value.id).unwrap(),
+            provider: Provider::from_str(&value.provider).unwrap(),
+            display_name: value.display_name,
+            username: value.username,
+            email: value.email,
+            enabled: value.enabled,
+            created_at: DateTime::from_str(&value.created_at).unwrap(),
+            last_used: value
+                .last_used
+                .map(|lu| DateTime::from_str(&lu).ok())
+                .unwrap(),
+            capabilities: value.capabilities.into_iter().map(Into::into).collect(),
         }
     }
 }
