+++ src/models/credentials.rs
@@ -1,0 +1,11 @@
+pub use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
+pub struct Credential {
+    pub access_token: String,
+    pub refresh_token: Option<String>,
+    pub expires_at: Option<DateTime<Utc>>,
+    pub scope: Vec<String>,
+    pub token_type: String,
+}
